Moved ColumnData component to its own file. (#1972)
