+++ imports/plugins/core/email/client/components/emailLogs.js
@@ -1,19 +1,17 @@
 import React, { Component, PropTypes } from "react";
-import moment from "moment";
-import { Card, CardHeader, CardBody, CardGroup, Icon, Loading } from "/imports/plugins/core/ui/client/components";
+import { Card, CardHeader, CardBody, CardGroup, Loading } from "/imports/plugins/core/ui/client/components";
 import MeteorGriddle from "/imports/plugins/core/ui-grid/client/griddle";
+import EmailTableColumn from "./emailTableColumn";
 import { Jobs } from "/lib/collections";
 import { i18next } from "/client/api";
 import "./emailConfig.css";
 
 
 class EmailLogs extends Component {
-
   renderEmailsTable() {
     const filteredFields = ["data.to", "updated", "data.subject", "status"];
     const filteredFieldsColumns = ["data.to", "updated", "data.subject", "status"];
     const noDataMessage = i18next.t("admin.logs.noEmails");
-
 
     // helper adds a class to every grid row
     const customRowMetaData = {
@@ -22,58 +20,13 @@ class EmailLogs extends Component {
       }
     };
 
-
-    const ColumnData = React.createClass({
-
-      handleAction() {
-        const emailId = this.props.rowData._id;
-        const emailAddress = this.props.rowData.data.to;
-
-        Meteor.call("emails/retryFailed", emailId, (err) => {
-          if (err) {
-            return Alerts.toast(i18next.t("app.error", { error: err.reason }), "error");
-          }
-          return Alerts.toast(i18next.t("mail.alerts.resendSuccess", { email: emailAddress }), "success");
-        });
-      },
-      render() {
-        const renderColumn = this.props.metadata.columnName;
-
-        if (renderColumn === "status") {
-          if (this.props.data === "completed") {
-            return (
-              <span>
-                <Icon icon="fa fa-circle" className="pull-left valid" />
-              </span>
-            );
-          }
-          return (
-            <span>
-              <Icon icon="fa fa-circle" className="pull-left error" />
-              <span onClick={this.handleAction} title={this.props.data}><Icon icon="fa fa-share" className="pull-right" /></span>
-            </span>
-          );
-        }
-        if (renderColumn === "updated") {
-          const createdDate = moment(this.props.data).format("LLL");
-          return (
-            <span>{createdDate}</span>
-          );
-        }
-        return (
-          <span>{this.props.data}</span>
-        );
-      }
-    });
-
-
     // add i18n handling to headers
     const customColumnMetadata = [];
     filteredFields.forEach(function (field) {
       const columnMeta = {
         columnName: field,
         displayName: i18next.t(`admin.logs.headers.${field}`),
-        customComponent: ColumnData
+        customComponent: EmailTableColumn
       };
       customColumnMetadata.push(columnMeta);
     });
